some fixes on 'long long int' matter
git-svn-id: file:///home/roberto/Git/svn@5699 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 4b03a64b949bb86b27c51e5eedfae03b7af14426

+++ trunk/xmipp/libraries/data/parallel_job_handler.h
@@ -53,23 +53,23 @@ private:
     int lockFile;
     bool fileCreator;
 
-    int numberOfJobs;
-    int blockSize;
-    int assignedJobs;
+    long long int numberOfJobs;
+    long long int blockSize;
+    long long int assignedJobs;
 
 public:
 
 
     //TODO: todos los casos: inicializar variable blocksize, crear base de datos,
-    ParallelJobHandler(int nJobs, int bSize, char *fName);
+    ParallelJobHandler(long long int nJobs, long long int bSize, char *fName);
     ParallelJobHandler(const char *fName);
     ~ParallelJobHandler();
 
 
-    bool setBlockSize(int blockSize);
+    bool setBlockSize(long long int blockSize);
     int getBlockSize();
 
-    bool getJobs(int &first, int &last); // False = no more jobs, true = more jobs
+    bool getJobs(long long int &first, long long int &last); // False = no more jobs, true = more jobs
